// src/api/openapi/model/workDay.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';
import { Expense } from './expense';
import { Drive } from './drive';
import { JobSection } from './jobSection';
import { AdditionalWorkload } from './additionalWorkload';
import { EatingPlace } from './eatingPlace';


export interface WorkDay { 
    additional_workloads: Array<AdditionalWorkload>;
    date: string;
    drives: Array<Drive>;
    eating_place?: EatingPlace | null;
    expenses: Array<Expense>;
    id: number;
    job_sections: Array<JobSection>;
    minutes: number;
    user: UserEssential;
}